(libT) (1) shave a few bytes from tr_peer and tr_ratecontrol's footprints. (2) remove some dead code.

+++ libtransmission/ratecontrol.c
@@ -42,8 +42,8 @@ struct tr_transfer
 
 struct tr_ratecontrol
 {
-    int limit;
-    int newest;
+    int8_t limit;
+    int8_t newest;
     struct tr_transfer transfers[HISTORY_SIZE];
 };
